cleaned

+++ app/Http/Controllers/Backend/StockEntryController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
-use App\Models\Backend\MedecineStock;
 use App\Models\Backend\ExpireDateMedecines;
 use App\Models\Backend\StockEntryLog;
 use App\Models\Backend\Manufacturer;
@@ -17,7 +16,6 @@ use App\Models\Backend\ProductType;
 use App\Models\Backend\ProductCategory;
 use App\Models\Backend\ProductSubCategory;
 use App\Models\Backend\MedecineUsage;
-use App\Models\Backend\Stock;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\DB;
 use Exception;
@@ -65,51 +63,51 @@ class StockEntryController extends Controller
         }
 
         $item_id = (int)$request->item_id;
-    try {
+        try {
             DB::beginTransaction();
-        $medecine = Product::find($item_id);
-        // return $medecine;
+            $medecine = Product::find($item_id);
+            // return $medecine;
 
-        $medecine->last_stock = $request->current_stock;
-        $medecine->current_stock = $request->current_stock;
-        $medecine->mrp_rate = $request->mrp_rate;
-        $medecine->tp_rate = $request->tp_rate;
-        $medecine->stock_location = $request->stock_location;
-        $medecine->stock_per = 100;
-        // $medecine->stock_date = $date;
-        $medecine->total_stock = $medecine->total_stock + $request->stock_quantity;
-        $medecine->save();
-        $item_exists_expiry = ExpireDateMedecines::where('medecine_id','=',$medecine->id)->where('expiry_date','=',$request->expire_date)->first();
-        if($item_exists_expiry){
-            $item_exists_expiry->stock_qty = (int)$item_exists_expiry->stock_qty + (int)$request->stock_quantity;
-            $item_exists_expiry->current_qty = (int)$item_exists_expiry->current_qty + (int)$request->stock_quantity;
-            $item_exists_expiry->save();
-        }else{
-            $expiry_wise = new ExpireDateMedecines();
-            $expiry_wise->medecine_id = $medecine->id;
-            $expiry_wise->expiry_date = $request->expire_date;
-            $expiry_wise->stock_qty = $request->stock_quantity;
-            $expiry_wise->current_qty = $request->stock_quantity;
-            $expiry_wise->save();
-        }
-        
-        $stock_log = new StockEntryLog();
-        $stock_log->medecine_id = $medecine->id;
-        $stock_log->mrr_id = $request->mrr_id;
-        $stock_log->stock_date = $date;
-        $stock_log->expiry_date = $request->expire_date;
-        $stock_log->manufacture_date = $request->manufacture_date;
-        $stock_log->stock_qty = $request->stock_quantity;
-        $stock_log->save();
+            $medecine->last_stock = $request->current_stock;
+            $medecine->current_stock = $request->current_stock;
+            $medecine->mrp_rate = $request->mrp_rate;
+            $medecine->tp_rate = $request->tp_rate;
+            $medecine->stock_location = $request->stock_location;
+            $medecine->stock_per = ceil(((int)$medecine->current_stock/(int)$medecine->last_stock)*100);
+            // $medecine->stock_date = $date;
+            $medecine->total_stock = $medecine->total_stock + $request->stock_quantity;
+            $medecine->save();
+            $item_exists_expiry = ExpireDateMedecines::where('medecine_id','=',$medecine->id)->where('expiry_date','=',$request->expire_date)->first();
+            if($item_exists_expiry){
+                $item_exists_expiry->stock_qty = (int)$item_exists_expiry->stock_qty + (int)$request->stock_quantity;
+                $item_exists_expiry->current_qty = (int)$item_exists_expiry->current_qty + (int)$request->stock_quantity;
+                $item_exists_expiry->save();
+            }else{
+                $expiry_wise = new ExpireDateMedecines();
+                $expiry_wise->medecine_id = $medecine->id;
+                $expiry_wise->expiry_date = $request->expire_date;
+                $expiry_wise->stock_qty = $request->stock_quantity;
+                $expiry_wise->current_qty = $request->stock_quantity;
+                $expiry_wise->save();
+            }
+            
+            $stock_log = new StockEntryLog();
+            $stock_log->medecine_id = $medecine->id;
+            $stock_log->mrr_id = $request->mrr_id;
+            $stock_log->stock_date = $date;
+            $stock_log->expiry_date = $request->expire_date;
+            $stock_log->manufacture_date = $request->manufacture_date;
+            $stock_log->stock_qty = $request->stock_quantity;
+            $stock_log->save();
 
-        DB::commit(); 
+            DB::commit(); 
 
-        return response()->json(['success'=>$medecine,'expiry_log'=>$stock_log]);
-        
-        } catch (Exception $e) {
-            DB::rollBack(); 
+            return response()->json(['success'=>$medecine,'expiry_log'=>$stock_log]);
+            
+            } catch (Exception $e) {
+                DB::rollBack(); 
 
-            return response()->json(['error' => 'Transaction failed.', 'message' => $e->getMessage()], 500);
+                return response()->json(['error' => 'Transaction failed.', 'message' => $e->getMessage()], 500);
         }
     }
 
@@ -118,7 +116,7 @@ class StockEntryController extends Controller
      */
     public function show(string $id)
     {
-        $stock = MedecineStock::find((int)$id);
+        $stock = Product::find((int)$id);
         $expiryDate = ExpireDateMedecines::where('medecine_id','=',(int)$id)
                         ->where('medecine_id','=',(int)$id)
                         ->where('current_qty','>',0)
@@ -128,60 +126,6 @@ class StockEntryController extends Controller
         return response()->json(['item'=>$stock,'expiryDates'=>$expiryDate]);
     }
 
-    public function todaySummary()
-    {
-
-        $date = Carbon::now()->format('Y-m-d');
-
-
-        $data['todaystock'] = StockEntryLog::leftJoin('medecine_stocks','stock_entry_logs.medecine_id','=','medecine_stocks.id')
-        ->select('stock_entry_logs.stock_qty','stock_entry_logs.stock_date','medecine_stocks.manufacturer','medecine_stocks.name','medecine_stocks.generic','medecine_stocks.strength','medecine_stocks.type','medecine_stocks.use_for','medecine_stocks.category','medecine_stocks.current_stock' )
-        ->where('stock_entry_logs.stock_date','=',$date)
-        ->get();
-        // return $data;
-        return view('backend.stockmedecine.todaystocksummary',$data);
-
-    }
-
-    public function stockEntrySummary()
-    {
-
-        $date = Carbon::now()->format('Y-m-d');
-
-
-        $data['todaystock'] = StockEntryLog::leftJoin('medecine_stocks','stock_entry_logs.medecine_id','=','medecine_stocks.id')
-        ->select('stock_entry_logs.stock_qty','stock_entry_logs.stock_date','medecine_stocks.manufacturer','medecine_stocks.name','medecine_stocks.generic','medecine_stocks.strength','medecine_stocks.type','medecine_stocks.use_for','medecine_stocks.category','medecine_stocks.current_stock' )
-        ->where('stock_entry_logs.stock_date','=',$date)
-        ->get();
-        // return $data;
-        return view('backend.stockmedecine.stocksummary',$data);
-
-    }
-
-    public function stockEntrySummaryDate(Request $request)
-    {
-
-
-        $validated = Validator::make($request->all(),[
-            'from_date' => 'required',
-            'to_date' => 'required',
-        ]);
-
-        if($validated->fails()){
-            return back()->withErrors($validated)->withInput();
-        }
-        $date = Carbon::now()->format('Y-m-d');
-        $date1 = $request->from_date;
-        $date2 = $request->to_date;
-
-        $data['todaystock'] = StockEntryLog::leftJoin('medecine_stocks','stock_entry_logs.medecine_id','=','medecine_stocks.id')
-        ->select('stock_entry_logs.stock_qty','stock_entry_logs.stock_date','medecine_stocks.manufacturer','medecine_stocks.name','medecine_stocks.generic','medecine_stocks.strength','medecine_stocks.type','medecine_stocks.use_for','medecine_stocks.category','medecine_stocks.current_stock' )
-        ->whereBetween('stock_entry_logs.stock_date',[$date1,$date2])
-        ->get();
-        // return $data;
-        return view('backend.stockmedecine.stocksummary',$data);
-
-    }
 
     /**
      * Show the form for editing the specified resource.
@@ -198,7 +142,6 @@ class StockEntryController extends Controller
     {
 
         try {
-            DB::beginTransaction(); 
 
             $stock_log = StockEntryLog::find($id);
 
@@ -208,11 +151,18 @@ class StockEntryController extends Controller
 
             $item_id = $stock_log->medecine_id;
             $medecine = Product::find($item_id);
+            if(((int)$expiry_wise->current_qty - (int)$stock_log->stock_qty) < 0){
+                return response()->json(['error'=>'Stock not available !!']);
+            }
+
+            DB::beginTransaction(); 
+
+            
             // return $medecine;
 
             $medecine->last_stock = ((int)$medecine->last_stock - (int)$stock_log->stock_qty)+(int)$request->stock_qty;;
             $medecine->current_stock = ((int)$medecine->current_stock - (int)$stock_log->stock_qty)+(int)$request->stock_qty;
-            $medecine->stock_per = 100;
+            $medecine->stock_per = ceil(((int)$medecine->current_stock/(int)$medecine->last_stock)*100);
             $medecine->total_stock = ((int)$medecine->total_stock - (int)$stock_log->stock_qty)+(int)$request->stock_qty;
             $medecine->save();
 
@@ -251,10 +201,15 @@ class StockEntryController extends Controller
         $medecine->last_stock = ((int)$medecine->last_stock - (int)$stock_log->stock_qty);
         $medecine->current_stock = ((int)$medecine->current_stock - (int)$stock_log->stock_qty);
         $medecine->total_stock = ((int)$medecine->total_stock - (int)$stock_log->stock_qty);
-        $medecine->save();
+        $medecine->stock_per = ceil(((int)$medecine->current_stock/(int)$medecine->last_stock)*100);
 
         $expiry_wise->stock_qty = ((int)$expiry_wise->stock_qty - (int)$stock_log->stock_qty);
         $expiry_wise->current_qty = ((int)$expiry_wise->current_qty - (int)$stock_log->stock_qty);
+        if(((int)$expiry_wise->current_qty - (int)$stock_log->stock_qty) < 0){
+            return response()->json(['error'=>'Stock not available !!']);
+        }
+        $medecine->save();
+
         $expiry_wise->update();
     
         $stock_log->delete();
@@ -269,15 +224,10 @@ class StockEntryController extends Controller
      * @param string $id
      * @return \Illuminate\Http\JsonResponse
      */
-    public function billingitems(string $id){
-       
-    }
-
-
 
     public function search(Request $request)
     {
-        $lastid = MedecineStock::where('name', 'like', '%'.$request->search.'%')->get();
+        $lastid = Product::where('name', 'like', '%'.$request->search.'%')->get();
         return $lastid;
     }
